+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 ## HEAD (Unreleased)
 * Upgrade to v2.6.0 of the Packet Terraform Provider
+* Add support for DotNet SDK Generation
 
 ---