+++ CHANGELOG.md
@@ -2,10 +2,13 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Upgrade to v2.6.1 of the Packet Terraform Provider
-* Add support for DotNet SDK Generation
+_(none)_
 
 ---
+
+## 1.1.0 (2019-11-15)
+* Upgrade to v2.6.1 of the Packet Terraform Provider
+* Add support for DotNet SDK Generation
 
 ## 1.0.0 (2019-10-18)
 * Regenerate SDK based on tf2pulumi 0.6.0